% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_id_mapping.R
\docType{data}
\name{users_id_mapping}
\alias{users_id_mapping}
\title{The mapping between behavior and fMRI user}
\format{
A named character vector with fMRI user id as names and behavior user
id as values.
}
\usage{
users_id_mapping
}
\description{
For some real-world reasons, the user id in behavior data and fMRI data are
not the same. This dataset provides the mapping between them.
}
\keyword{datasets}
